+++ lib/mcpwm/motor.h
@@ -4,7 +4,7 @@
 class motor {
    public:
     motor(PinName pin, mcpwm_unit_t mcpwm_num, mcpwm_timer_t timer_num, mcpwm_io_signals_t io_signal, PinName dir0, PinName dir1);
-    void write(int duty);
+    void write(float duty);
 
    private:
     gpio dir0;
